+++ f/frizzle/kobotoolbox/README.md
@@ -1,4 +1,4 @@
-# Kobo: Fetch Survey Responses
+# KoboToolbox: Fetch Survey Responses
 
 This script fetches form submissions from the KoboToolbox REST API, transforms the data for SQL compatibility, and stores it in a PostgreSQL database. Additionally, it downloads any attachments and saves them to a specified directory.
