+++ f/frizzle/kobotoolbox/kobotoolbox_responses.script.yaml
@@ -1,4 +1,4 @@
-summary: 'Kobo: Fetch Survey Responses'
+summary: 'KoboToolbox: Fetch Survey Responses'
 description: Uses KoboToolbox REST API to download all form submissions
 lock: '!inline f/frizzle/kobo/kobo_responses.script.lock'
 concurrency_time_window_s: 0
