+++ f/frizzle/kobotoolbox/tests/assets/trees.png
