+++ tox.ini
@@ -1,14 +1,14 @@
 [tox]
 skipsdist = True
-env_list = kobo
+env_list = kobotoolbox
 
 [testenv]
 setenv =
     PYTHONPATH=.
 
-[testenv:kobo]
+[testenv:kobotoolbox]
 deps =
-    -r{toxinidir}/f/frizzle/kobo/kobo_responses.script.lock
-    -r{toxinidir}/f/frizzle/kobo/tests/requirements-test.txt
+    -r{toxinidir}/f/frizzle/kobotoolbox/kobotoolbox_responses.script.lock
+    -r{toxinidir}/f/frizzle/kobotoolbox/tests/requirements-test.txt
 commands =
-    pytest {posargs} f/frizzle/kobo
+    pytest {posargs} f/frizzle/kobotoolbox
